backend/requirements.txt
# Tessera Backend Requirements

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async file handling
aiofiles==23.2.1

# Data handling
h5py==3.10.0
numpy==1.26.3

# Machine learning
scikit-learn==1.4.0
umap-learn==0.5.5

# Data validation
pydantic==2.5.3

# HTTP client (for CLI)
httpx==0.26.0

# Development
python-dotenv==1.0.0
